requirements.txt
google-adk>=1.28.0,<2.0.0
fastapi>=0.116.0,<1.0.0
